refactor:group CP modules under fla/cp/
